stream size was not used for size comparison
This results in some cases in suboptimal results.

e.g.
stream1 + lut is smaller than best_stream by <4 bytes
therefore it is selected
next iteration: the "best" stream is now larger than the previous best stream.
In case of HC, this happened a few times for 1 subsequence hence resulting in a stream that is a few bytes larger than the actual smallest one