added script for parallel processing of perform_codec_comparison
This scrip takes 2 inputs:
A path to the folder with files
A number indicating the number of parallel processes

Note: It is possible that the number of processes running is 1 or 2 higher at some point, but it will return to the maximum number. This is due to the way the number of processes is derived, however this is the most reliable approach I found until now.
note2: you can multiple of these scripts in parallel on one node/computer, however the number of running treads will be equal to max_nr_of_threads across all scripts.